Moves weather and location into App.jsx in order to update
